need to work on cart and whoami

+++ src/components/Signup.jsx
@@ -11,7 +11,8 @@ export default function Signup({setUser}) {
       password: "",
     }),
     [pwConfirmation, setPwConfirmation] = useState(""),
-    history = useHistory();
+    history = useHistory(),
+    [error, setError] = useState(false);
 
 
     const [form, setForm] = React.useState({
@@ -28,11 +29,15 @@ export default function Signup({setUser}) {
   
   function handleSubmit(e) {
     e.preventDefault();
-    signup({ user: { ...form } }).then((json) => {
-        localStorage.setItem('token', json.jwt)
-        setUser(json.user)
-      history.push("/profile");
-    });
+    if(form.password !== pwConfirmation) {
+      setError(true)
+    } else {
+      signup({ user: { ...form } }).then((json) => {
+          localStorage.setItem('token', json.jwt)
+          setUser(json.user)
+        history.push("/profile");
+      })
+    }
   }
   
   return (
@@ -75,11 +80,15 @@ export default function Signup({setUser}) {
             placeholder="Re-enter Your Password"
             value={pwConfirmation}
             name="password_confirmation"
-            onChange={(e) => setPwConfirmation(e.target.value)}
+            onChange={(e) => {
+              setError(false)
+              setPwConfirmation(e.target.value)}
+            }
           />
         </label>
         <button>Submit</button>
       </form>
+      { error && 'Passwords do not match'}
     </div>
   );
 }
